requirements.txt
aiohttp==3.6.2
nltk==3.4.5
matplotlib==3.1.3
Click==7.0
telepot==12.7
gevent_socketio==0.3.6
mysql_connector_repackaged==0.3.1
rasa==1.8.2
sqlparse==0.3.1
